+++ karma.conf.js
@@ -39,7 +39,7 @@ module.exports = function(config) {
 
     // list of files / patterns to exclude
     exclude: [
-      'spec/Content*Spec.js'
+      'spec/*ZipSpec.js',
     ],